refactor: App클래스 분리
- index.js에 구현한 App 클래스를 App.js 폴더로 이동
